refactor(workflow-tasks): move ComponentRegistry into library
